channels.scm: add emacs-master
(this e-mail was written with vc-prepare-patch)

Greetings,

At the end of this e-mail is a diff for adding my channel, emacs-master.

It's a fork of both guix-channel-emacs-master[1] and Divya-lambda[2],
with the difference coming from the fact that the commit is updated
automatically with a GitHub Actions workflow (a similar process is also
found on guix-channel-emacs-master, though not using GitHub
as it's hosted on Codeberg.). Despite the automation, the channel is signed.

Akib from guix-channel-emacs-master said they don't plan to maintain
their channel[3], so I thought that sending a patch here would be best.

(And there's also a bug not addressed in the guix patches which I fixed)

An alternative to this would be to introduce the same workflow to
divya-lambda. The codeberg repo of divya-lambda could then be synced
with GitHub through a webhook[4] (I think so, never messed with those).

[1] Azmain Turja, A. (2023) “guix-channel-emacs-master.” Available at:
https://codeberg.org/akib/guix-channel-emacs-master (Accessed:
February 03, 2025).
[2] Ranjan, D. (2024) “Divya-lambda.” Available at:
https://codeberg.org/divyaranjan/divya-lambda (Accessed:
February 03, 2025).
[3]
https://github.com/gs-101/emacs-master/issues/1#issuecomment-2612809453
[4]
https://forgejo.org/docs/latest/user/webhooks/

Regards,

--
Gabriel Santos